+++ __tests__/index.ts
@@ -1,2 +1,2 @@
-export * as cloudRun from "./cloudrun/index"
-export * as compute from "./compute/index"
+export * as cloudRun from './cloudrun/index'
+export * as compute from './compute/index'