+++ __tests__/cloudrun/index.ts
@@ -1,1 +1,1 @@
-export * as cloudrunService from "./cloudrunService"
+export * as cloudrunService from "./disallowPublicIngress"