+++ __tests__/cloudrunv2/index.ts
@@ -1,1 +1,1 @@
-export * as cloudrunV2Service from "../cloudrunv2/cloudrunv2Service"
+export * as cloudrunV2Service from "./disallowPublicIngress"