+++ src/cloudrunv2/disallowPublicIngress.ts
@@ -1,6 +1,6 @@
 import { ResourceValidationArgs, ReportViolation, EnforcementLevel } from "@pulumi/policy";
 
-export const cloudrunv2DisallowPublicIngress = {
+export const disallowPublicIngress = {
 	name: "cloudrunv2-disallow-public-ingress",
 	description: "Check that CloudRun2 services do not have public ingress set to 'all'.",
 	enforcementLevel: "advisory" as EnforcementLevel,