+++ src/cloudfunctions/index.ts
@@ -1,11 +1,13 @@
 // import { disallowEndOfLifeRuntime } from "./disallowEndOfLifeRuntime";
 import { disallowEnvsSecrets } from "./disallowEnvsSecrets";
+import { disallowPlainHttp } from "./disallowPlainHttp";
 import { disallowPublicIngress } from "./disallowPublicIngress";
-// import { requireCmek } from "./requireCmek";
+import { requireCmek } from "./requireCmek";
 
 export const cloudfunctionsPolicies = [
     // disallowEndOfLifeRuntime,
 	disallowEnvsSecrets,
+	disallowPlainHttp,
 	disallowPublicIngress,
-    // requireCmek,
+    requireCmek,
 ];