+++ __tests__/index.ts
@@ -1,3 +1,5 @@
+export * as cloudfunctions from './cloudfunctions'
+export * as cloudfunctionsv2 from './cloudfunctionsv2'
 export * as cloudrun from './cloudrun'
 export * as cloudrunv2 from './cloudrunv2'
 export * as compute from './compute'