+++ src/cloudfunctions/disallowPlainHttp.ts
@@ -1,0 +1,18 @@
+import { ResourceValidationArgs, ReportViolation, EnforcementLevel } from "@pulumi/policy";
+
+export const disallowPlainHttp = {
+	name: "cloudrunfunctions-function-disallow-plain-http",
+	description: "Check that CloudFunctions function don't accept plain 'HTTP'.",
+	enforcementLevel: "advisory" as EnforcementLevel,
+	validateResource: (args: ResourceValidationArgs, reportViolation: ReportViolation) => {
+		if (args.type === "gcp:cloudfunctions/function:Function") {
+			const trigger = args.props.triggerHttp;
+			const httpsTrigger = args.props.httpsTriggerSecurityLevel;
+            if (trigger && httpsTrigger === "SECURE_OPTIONAL") {
+                reportViolation(
+                    "CloudFunctions function shouldn't accept plain 'HTTP' connections. Consider using 'SECURE_ALWAYS'."
+                );
+            }
+		}
+	},
+}