+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add CodeQL analysis (#4).
+- Add dependabot npm scan (#4).
 
 ## [1.1.0] - 2023-12-25