+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add CodeQL analysis (#4).
+- Add dependabot npm scan (#4).
+
 ## [1.1.0] - 2023-12-25
 
 ### Added