+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added `.editorconfig` file (#8).
+- CloudFunctions 1st and 2nd generation policies (#8):
+  - `disallowEnvsSecrets`
+  - `disallowPublicIngress`
+
 ## [1.1.1] - 2023-12-26
 
 ### Added