+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `.editorconfig` file (#8).
 - CloudFunctions 1st and 2nd generation policies (#8):
   - `disallowEnvsSecrets`
+  - `disallowPlainHttp`
   - `disallowPublicIngress`
+  - `requireCmek`
 
 ## [1.1.1] - 2023-12-26