+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `disallowEnvsSecrets`
   - `disallowPlainHttp`
   - `disallowPublicIngress`
+  - `disallowVpcConnectorPublicEgress`
   - `requireCmek`
 
 ## [1.1.1] - 2023-12-26